asdasd

+++ Inc/signal.h
@@ -16,7 +16,7 @@ extern char dshot;
 extern char servoPwm;
 extern uint32_t gcr[];
 extern int armed_count_threshold;
-extern uint16_t degrees_celsius;
+extern uint8_t degrees_celsius;
 extern char crawler_mode;
 
 extern uint16_t ADC_raw_volts;
